Move DedupAddrs test
